requirements.txt
streamlit>=1.28.0
chromadb>=0.4.15
openai>=1.3.0
python-dotenv>=1.0.0
pypdf2>=3.0.1
python-docx>=1.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
numpy>=1.24.0
sentence-transformers>=2.2.0

